Extract fraud_proof module in sp-domains
No logical changes.
