+++ crates/pallet-domains/src/lib.rs
@@ -28,10 +28,8 @@ use frame_system::offchain::SubmitTransaction;
 pub use pallet::*;
 use sp_core::H256;
 use sp_domains::bundle_election::{verify_system_bundle_solution, verify_vrf_proof};
-use sp_domains::{
-    BundleEquivocationProof, ExecutionReceipt, FraudProof, InvalidTransactionCode,
-    InvalidTransactionProof, ProofOfElection, SignedOpaqueBundle,
-};
+use sp_domains::fraud_proof::{BundleEquivocationProof, FraudProof, InvalidTransactionProof};
+use sp_domains::{ExecutionReceipt, InvalidTransactionCode, ProofOfElection, SignedOpaqueBundle};
 use sp_runtime::traits::{BlockNumberProvider, CheckedSub, One, Saturating, Zero};
 use sp_runtime::transaction_validity::TransactionValidityError;
 use sp_runtime::RuntimeAppPublic;
@@ -43,9 +41,9 @@ mod pallet {
     use frame_support::PalletError;
     use frame_system::pallet_prelude::*;
     use sp_core::H256;
+    use sp_domains::fraud_proof::{BundleEquivocationProof, FraudProof, InvalidTransactionProof};
     use sp_domains::{
-        BundleEquivocationProof, ExecutionReceipt, ExecutorPublicKey, FraudProof,
-        InvalidTransactionCode, InvalidTransactionProof, SignedOpaqueBundle,
+        ExecutionReceipt, ExecutorPublicKey, InvalidTransactionCode, SignedOpaqueBundle,
     };
     use sp_runtime::traits::{
         CheckEqual, MaybeDisplay, MaybeMallocSizeOf, One, SimpleBitOps, Zero,
