+++ domains/client/domain-executor/src/fraud_proof.rs
@@ -7,7 +7,8 @@ use sp_api::ProvideRuntimeApi;
 use sp_blockchain::HeaderBackend;
 use sp_core::traits::{CodeExecutor, SpawnNamed};
 use sp_core::H256;
-use sp_domains::{ExecutionPhase, ExecutionReceipt, FraudProof};
+use sp_domains::fraud_proof::{ExecutionPhase, FraudProof};
+use sp_domains::ExecutionReceipt;
 use sp_runtime::generic::BlockId;
 use sp_runtime::traits::{Block as BlockT, HashFor, Header as HeaderT, NumberFor};
 use sp_trie::StorageProof;
